Digitize op nits

+++ keras_core/backend/tensorflow/numpy.py
@@ -229,7 +229,7 @@ def diagonal(x, offset=0, axis1=0, axis2=1):
 def digitize(x, bins):
     x = convert_to_tensor(x)
     bins = list(bins)
-    return tf.cast(tf.raw_ops.Bucketize(input=x, boundaries=bins), "int64")
+    return tf.raw_ops.Bucketize(input=x, boundaries=bins)
 
 
 def dot(x, y):
